Better management of metadata exchange
When the peer that has the torrent sends a bitfield and we don't have
the torrent, we must store the other's bitfield until we finish the
metadata exchange, then try to set this "temp" value. Otherwise we
naively say it has no piece, which may be wrong.

Also shortcut some straightforward paths