Use a long lived connection + smarter detection of new content
Long lived connections aren't closed when peers call close; they are
parked to be later reused with the same peer.

If the connection isn't reused for the next 1 minute, the underlying
TCP connection is closed (but this process is invisible to callers)

Also, when the watcher or the controlSession receives information
about a new torrent, only switch to it if it is really different
than the one we are currently on.