Use CouchDB-like revision ids
We use revision ids that look like

    2-5a4dada2423a45d729c40e76f08229e449136854e4

to have some sense of history. In the future we will use it to know
what version of the data we should keep.

We also have to serialize this data in the "current" control file,
so we can start from it next time we run the application. We also
serialize the (future) signature in this file. The format is exactly
the format of the ihmessage we send to other peers.