Wait for a stop in changes to update current meta
In case the changes take more than 10 seconds, some changes will
occur while the meta is recalculated, so the meta will be invalid
which will only be known when trying to launch the current session.

We now wait for one run where there are no changes to start creating
the meta.