Compartmentalize commands with codegangsta/cli
To generate a share, use

        $ rakoshare gen -dir <the dir>

This will generate a 3-tuple of ids and save this tuple with the dir in
a sql db.

To share the newly generated content, use

        $ rakoshare share -id <one of the ids>

The first lines will indicate what "mode" you're in. For example, here
you can Read and Store data but you can't Write:

WriteReadStore:
     ReadStore: rxJwfAAANNukRDNGfEdWuNQEYXEmvLKmwhYK2rqC5i5K
         Store: 2PYGgFGrMyFVjc2wrJ3uS8Ed41yXgFHmRXjbeUCXehs7u

To enter a share after receiving a shareid, use

        $ rakoshare share -id <the share id> -dir <where to store data>