Solve peer connection race lazily
Before we discarded all connection attempts that matched an existing
peer we already had. This has issues when 2 peers simultanously try
to connect to each other, because each half will be cut.

Now we accept all connections, and gradually kill duplicates during
keepalives. By making it at this time we are sure that connections
are completely established and it is safe to cut them.